pyproject.toml
[tool.poetry]
name = "core-supply"
version = "0.1.0"
description = ""
authors = ["Leandro Herdy <le.herdy@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
pandas = "^2.2.3"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pytest = "^8.3.3"
pytest-xdist = "^3.6.1"
pytest-django = "^4.9.0"
model-bakery = "^1.20.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["core/clients/tests"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"
pythonpath = "."

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
